refactor(dist): privatize imports from `dist::dist`
