+++ src/cli/common.rs
@@ -16,8 +16,7 @@ use tracing::{debug, error, info, trace, warn};
 use super::self_update;
 use crate::cli::download_tracker::DownloadTracker;
 use crate::currentprocess::{terminalsource, Process};
-use crate::dist::dist::{TargetTriple, ToolchainDesc};
-use crate::dist::manifest::ComponentStatus;
+use crate::dist::{manifest::ComponentStatus, TargetTriple, ToolchainDesc};
 use crate::install::UpdateStatus;
 use crate::toolchain::names::{LocalToolchainName, ToolchainName};
 use crate::toolchain::toolchain::Toolchain;
