+++ src/cli/self_update/windows.rs
@@ -15,7 +15,7 @@ use super::common;
 use super::{install_bins, InstallOpts};
 use crate::cli::download_tracker::DownloadTracker;
 use crate::currentprocess::Process;
-use crate::dist::dist::TargetTriple;
+use crate::dist::TargetTriple;
 use crate::utils::utils;
 use crate::utils::Notification;
