+++ tests/suite/dist_install.rs
@@ -5,10 +5,10 @@ use rustup::currentprocess::TestProcess;
 use rustup::dist::component::Components;
 use rustup::dist::component::Transaction;
 use rustup::dist::component::{DirectoryPackage, Package};
-use rustup::dist::dist::DEFAULT_DIST_SERVER;
 use rustup::dist::prefix::InstallPrefix;
 use rustup::dist::temp;
 use rustup::dist::Notification;
+use rustup::dist::DEFAULT_DIST_SERVER;
 use rustup::utils::utils;
 
 use rustup::test::mock::{MockComponentBuilder, MockFile, MockInstallerBuilder};
